+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Expanded `IntervalBarSeries` and `TornadoBarSeries` to allow for varied label positions and angles (#2027)
 - VectorSeries (#107)
 - LogarithmicColorAxis (#92)
+- DateTimeAxis.DateTimePrecision property (related to #2061)
 
 ### Changed
 - Make consistent BaseValue and BaseLine across BarSeries, LinearBarSeries, and HistogramSeries
@@ -26,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - Add support for .NET 7.0 (#1937)
 - Update SkiaSharp to Version 2.88.6
 - AxisRendererBase is now generic
+- DateTimeAxis.ToDateTime(double value) is now obsolete, replacements are provided (related to #2061)
 
 ### Removed
 - Support for .NET Framework 4.0 and 4.5 (#1839)
@@ -41,6 +43,7 @@ All notable changes to this project will be documented in this file.
 - Font weight not being applied in ImageSharp (#2006)
 - SkiaSharp - Fix use of obsolete functions (#1937)
 - Dashed lines are solid when exporting via SkiaSharp.SvgExporter (#1674)
+- DateTimeAxis.ToDateTime doesn't behave as intended in .NET 8 (#2061) 
 
 ## [2.1.2] - 2022-12-03